Update patch set 4
Patch Set 4:

(10 comments)

some suggestions for improvements but my main issue is in a comment below re. the idea that each template is bound to a certain schema_set?! I don't think that is correct, was that discussed with Tony?

I did not review the testware until this is clarified...

Patch-set: 4
Reviewer: Gerrit User 6639 <6639@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>