Update patch set 4
Patch Set 4:

(2 comments)

> Patch Set 4:
> 
> (10 comments)
> 
> some suggestions for improvements but my main issue is in a comment below re. the idea that each template is bound to a certain schema_set?! I don't think that is correct, was that discussed with Tony?
> 
> I did not review the testware until this is clarified...

The table which we have now in the app will have an entry for a single xpath query. If we look at that way, it will be bound to a single schema set. We could remove it from the id(currently it is a composite key, id + schema-set) and have only id as a primary key, but still, the template will be bound to a single schema set.

We could handle the multiple queries in a layer on top of this one, may be using a different table

Patch-set: 4
Reviewer: Gerrit User 2671 <2671@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>